+++ CDF3/src/libf/inq_vartypef.c
@@ -21,7 +21,7 @@
 
 /* Prototypes for the Fortran interfaces */
 #include "mpifnetcdf.h"
-FORTRAN_API int FORT_CALL nfmpi_inq_vartype_ ( int *v1, int64_t *v2, int *v3 ){
+FORTRAN_API int FORT_CALL nfmpi_inq_vartype_ ( int *v1, int *v2, int *v3 ){
     int ierr;
     int l2 = *v2 - 1;
     ierr = ncmpi_inq_vartype( *v1, l2, (nc_type *)(v3) );
